Broke big function up into two smaller functions to better structure the code
